requirements.txt
black~=21.12b0
boto3==1.24.88
botocore==1.27.88
certifi~=2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
idna==3.4
jmespath==1.0.1
mypy-extensions==0.4.3
pathspec~=0.9.0
platformdirs~=2.4.1
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
tomli~=1.2.3
typing-extensions==4.4.0
urllib3==1.26.12

pip~=20.2.1
py~=1.11.0
lxml~=4.9.1
aiohttp~=3.8.3
chardet~=3.0.4
pytest~=2.9.1
setuptools~=49.2.1
python-dotenv~=0.19.2
fastapi~=0.85.0